app-python/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
jinja2==3.1.4
python-multipart==0.0.20
httpx==0.28.1
google-generativeai==0.8.3
pdfminer.six==20231228
python-dotenv==1.0.1
